Moves closing a window to another module, refactoring the code.
